flatten file structure
