+++ packages/reacord/library/components/button.tsx
@@ -1,12 +1,9 @@
 import type { APIMessageComponentButtonInteraction } from "discord.js"
 import { randomUUID } from "node:crypto"
 import React from "react"
-import { ReacordElement } from "../../internal/element.js"
-import type { ComponentInteraction } from "../../internal/interaction"
-import type { MessageOptions } from "../../internal/message"
-import { getNextActionRow } from "../../internal/message"
-import { Node } from "../../internal/node.js"
-import type { ComponentEvent } from "../component-event"
+import type { ComponentEvent } from "../core/component-event.js"
+import { ReacordElement } from "../internal/element.js"
+import { Node } from "../node.js"
 import type { ButtonSharedProps } from "./button-shared-props"
 
 /**
@@ -54,31 +51,31 @@ export class ButtonNode extends Node<ButtonProps> {
 
   // this has text children, but buttons themselves shouldn't yield text
   // eslint-disable-next-line class-methods-use-this
-  override get text() {
-    return ""
-  }
+  // override get text() {
+  //   return ""
+  // }
 
-  override modifyMessageOptions(options: MessageOptions): void {
-    getNextActionRow(options).push({
-      type: "button",
-      customId: this.customId,
-      style: this.props.style ?? "secondary",
-      disabled: this.props.disabled,
-      emoji: this.props.emoji,
-      label: this.children.findType(ButtonLabelNode)?.text,
-    })
-  }
+  // override modifyMessageOptions(options: MessageOptions): void {
+  //   getNextActionRow(options).push({
+  //     type: "button",
+  //     customId: this.customId,
+  //     style: this.props.style ?? "secondary",
+  //     disabled: this.props.disabled,
+  //     emoji: this.props.emoji,
+  //     label: this.children.findType(ButtonLabelNode)?.text,
+  //   })
+  // }
 
-  override handleComponentInteraction(interaction: ComponentInteraction) {
-    if (
-      interaction.type === "button" &&
-      interaction.customId === this.customId
-    ) {
-      this.props.onClick(interaction.event)
-      return true
-    }
-    return false
-  }
+  // override handleComponentInteraction(interaction: ComponentInteraction) {
+  //   if (
+  //     interaction.type === "button" &&
+  //     interaction.customId === this.customId
+  //   ) {
+  //     this.props.onClick(interaction.event)
+  //     return true
+  //   }
+  //   return false
+  // }
 }
 
 class ButtonLabelNode extends Node<{}> {}
