+++ packages/reacord/library/components/embed-child.ts
@@ -1,4 +1,4 @@
-import { Node } from "../../internal/node.js"
+import { Node } from "../internal/node.js"
 import type { EmbedOptions } from "./embed-options"
 
 export abstract class EmbedChildNode<Props> extends Node<Props> {
