+++ packages/reacord/library/components/embed-image.tsx
@@ -1,29 +1,29 @@
 import React from "react"
-import { ReacordElement } from "../../internal/element.js"
+import { ReacordElement } from "../internal/element.js"
 import { EmbedChildNode } from "./embed-child.js"
 import type { EmbedOptions } from "./embed-options"
 
 /**
  * @category Embed
  */
-export type EmbedThumbnailProps = {
+export type EmbedImageProps = {
   url: string
 }
 
 /**
  * @category Embed
  */
-export function EmbedThumbnail(props: EmbedThumbnailProps) {
+export function EmbedImage(props: EmbedImageProps) {
   return (
     <ReacordElement
       props={props}
-      createNode={() => new EmbedThumbnailNode(props)}
+      createNode={() => new EmbedImageNode(props)}
     />
   )
 }
 
-class EmbedThumbnailNode extends EmbedChildNode<EmbedThumbnailProps> {
+class EmbedImageNode extends EmbedChildNode<EmbedImageProps> {
   override modifyEmbedOptions(options: EmbedOptions): void {
-    options.thumbnail = { url: this.props.url }
+    options.image = { url: this.props.url }
   }
 }
