+++ packages/reacord/library/components/embed-options.ts
