+++ packages/reacord/library/components/embed-thumbnail.tsx
@@ -1,29 +1,29 @@
 import React from "react"
-import { ReacordElement } from "../../internal/element.js"
+import { ReacordElement } from "../internal/element.js"
 import { EmbedChildNode } from "./embed-child.js"
 import type { EmbedOptions } from "./embed-options"
 
 /**
  * @category Embed
  */
-export type EmbedImageProps = {
+export type EmbedThumbnailProps = {
   url: string
 }
 
 /**
  * @category Embed
  */
-export function EmbedImage(props: EmbedImageProps) {
+export function EmbedThumbnail(props: EmbedThumbnailProps) {
   return (
     <ReacordElement
       props={props}
-      createNode={() => new EmbedImageNode(props)}
+      createNode={() => new EmbedThumbnailNode(props)}
     />
   )
 }
 
-class EmbedImageNode extends EmbedChildNode<EmbedImageProps> {
+class EmbedThumbnailNode extends EmbedChildNode<EmbedThumbnailProps> {
   override modifyEmbedOptions(options: EmbedOptions): void {
-    options.image = { url: this.props.url }
+    options.thumbnail = { url: this.props.url }
   }
 }
