+++ packages/reacord/library/text-node.ts
